refactor(mcp): modularize filters/format and add shared testable utils
