+++ crates/mcp/Cargo.toml
@@ -20,3 +20,5 @@ tracing = "0.1"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["test-util", "macros"] }
+insta = "1"
+proptest = "1"
