+++ crates/mcp/src/domain/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod documents_input;
 pub(crate) mod notes_input;
 pub(crate) mod search_input;
 pub(crate) mod tags_input;
