+++ crates/mcp/src/domain/tags_input.rs
@@ -1,9 +1,34 @@
 use crate::schema::{EditTagsRequest, ListTagsRequest};
 
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum TagListDetail {
+    Names,
+    Counts,
+    Full,
+}
+
+impl TagListDetail {
+    pub(crate) fn from_request(detail: Option<&str>) -> Self {
+        match detail {
+            Some("names") => Self::Names,
+            Some("full") => Self::Full,
+            _ => Self::Counts,
+        }
+    }
+
+    pub(crate) fn as_str(self) -> &'static str {
+        match self {
+            Self::Names => "names",
+            Self::Counts => "counts",
+            Self::Full => "full",
+        }
+    }
+}
+
 #[derive(Debug, Clone)]
 pub(crate) struct ListTagsInput {
     pub folder: Option<String>,
-    pub detail: String,
+    pub detail: TagListDetail,
     pub recursive: bool,
 }
 
@@ -22,7 +47,7 @@ pub(crate) enum TagsInputError {
 pub(crate) fn prepare_list_tags_input(req: ListTagsRequest) -> ListTagsInput {
     ListTagsInput {
         folder: req.folder,
-        detail: req.detail.unwrap_or_else(|| "counts".to_string()),
+        detail: TagListDetail::from_request(req.detail.as_deref()),
         recursive: req.recursive.unwrap_or(true),
     }
 }
@@ -54,8 +79,31 @@ mod tests {
             recursive: None,
             detail: None,
         });
-        assert_eq!(input.detail, "counts");
+        assert_eq!(input.detail, TagListDetail::Counts);
         assert!(input.recursive);
+    }
+
+    #[test]
+    fn list_tags_parses_all_detail_modes() {
+        let names = prepare_list_tags_input(ListTagsRequest {
+            folder: None,
+            recursive: None,
+            detail: Some("names".to_string()),
+        });
+        let full = prepare_list_tags_input(ListTagsRequest {
+            folder: None,
+            recursive: None,
+            detail: Some("full".to_string()),
+        });
+        let unknown = prepare_list_tags_input(ListTagsRequest {
+            folder: None,
+            recursive: None,
+            detail: Some("something-new".to_string()),
+        });
+
+        assert_eq!(names.detail, TagListDetail::Names);
+        assert_eq!(full.detail, TagListDetail::Full);
+        assert_eq!(unknown.detail, TagListDetail::Counts);
     }
 
     #[test]
