+++ crates/mcp/src/lib.rs
@@ -13,10 +13,10 @@ mod ports;
 mod schema;
 mod use_cases;
 
+pub use date_parser::{DateBound, parse_date};
 pub use format::{
     format_create_result, format_edit_result, format_edit_tags_result, format_entries,
-    format_examine_result, format_list_tags, format_results, format_tree_too_large,
-    format_vault_tree,
+    format_results, format_tree_too_large, format_vault_tree,
 };
 pub use schema::{
     CreateNoteRequest, EditNoteRequest, EditTagsRequest, ExamineRequest, ListTagsRequest,
